task: delete previous product records from today before inserting new
Haven't tested because the servers are down right now and I'm too
lazy to generate fake data. The reason why I am doing this is
because saving price records for the same product multiple times
didn't make sense. We are mostly interested with average prices so
storing data from each hour is useless.